+++ tests/testthat/test_calc_CosmicDoseRate.R
@@ -3,26 +3,44 @@ test_that("input validation", {
 
   expect_error(calc_CosmicDoseRate(depth = "error"),
                "'depth' should be of class 'numeric'")
+  expect_error(calc_CosmicDoseRate(depth = numeric(0), density = 1.7,
+                                   latitude = 38.1, longitude = 1.4),
+               "'depth' cannot be an empty numeric")
   expect_error(calc_CosmicDoseRate(depth = -2, density = 1.7, altitude = 364,
                                    latitude = 38.1, longitude = 1.4),
                "No negative values allowed for 'depth' and 'density'")
+  expect_error(calc_CosmicDoseRate(depth = 10, density = "error"),
+               "'density' should be of class 'numeric'")
+  expect_error(calc_CosmicDoseRate(depth = 10, density = numeric(0),
+                                   latitude = 38.1, longitude = 1.4),
+               "'density' cannot be an empty numeric")
   expect_error(calc_CosmicDoseRate(depth = 2.78, density = 1.7, altitude = 364,
                                    latitude = 38.1, longitude = 1.4,
                                    corr.fieldChanges = TRUE),
                "requires an age estimate")
   expect_error(calc_CosmicDoseRate(depth = 2.78, density = 1.7,
-                                   corr.fieldChanges = TRUE, est.age = 20,
                                    latitude = 38.06451),
                "'longitude' should be of class 'numeric'")
   expect_error(calc_CosmicDoseRate(depth = 2.78, density = 1.7,
-                                   corr.fieldChanges = TRUE, est.age = 20,
+                                   latitude = 38.1, longitude = numeric(0)),
+               "'longitude' cannot be an empty numeric")
+  expect_error(calc_CosmicDoseRate(depth = 2.78, density = 1.7,
                                    latitude = 38.06451, longitude = 1.49646),
                "'altitude' should be of class 'numeric'")
+  expect_error(calc_CosmicDoseRate(depth = 2.78, density = 1.7,
+                                   latitude = 38.06451, longitude = 1.49646,
+                                   altitude = numeric(0)),
+               "'altitude' cannot be an empty numeric")
   expect_error(calc_CosmicDoseRate(depth = 2.78, density = c(1.7, 2.9),
                                    corr.fieldChanges = TRUE, est.age = 20,
                                    latitude = 38.06451, longitude = 1.49646,
                                    altitude = 364),
-               "If you provide more than one value for density")
+               "The number of values for 'density' should either be 1 or")
+  expect_error(calc_CosmicDoseRate(depth = rep(2.78, 3), density = c(1.7, 2.9),
+                                   latitude = 38.06451, longitude = 1.49646,
+                                   altitude = 364),
+               "The number of values for 'density' should either be 1 or")
+
   expect_output(calc_CosmicDoseRate(depth = 2.78, density = 1.7,
                                     corr.fieldChanges = TRUE, est.age = 100,
                                     latitude = 38.0645, longitude = 1.4964,